tentando melhorar atividade
